+++ python/cugraph/traversal/sssp.py
@@ -14,28 +14,95 @@
 import numpy as np
 
 import cudf
-from cugraph.utilities import ensure_cugraph_obj
 from cugraph.structure import Graph, DiGraph
 from cugraph.traversal import sssp_wrapper
+from cugraph.utilities import (ensure_cugraph_obj,
+                               is_matrix_type,
+                               is_cp_matrix_type,
+                               import_optional,
+                               )
 
 # optional dependencies used for handling different input types
-try:
-    import cupy as cp
-    from cupyx.scipy.sparse.coo import coo_matrix as cp_coo_matrix
-    from cupyx.scipy.sparse.csr import csr_matrix as cp_csr_matrix
-    from cupyx.scipy.sparse.csc import csc_matrix as cp_csc_matrix
-except ModuleNotFoundError:
-    cp = None
-try:
-    import networkx as nx
-except ModuleNotFoundError:
-    nx = None
+nx = import_optional("networkx")
+
+cp = import_optional("cupy")
+cp_coo_matrix = import_optional("coo_matrix",
+                                import_from="cupyx.scipy.sparse.coo")
+cp_csr_matrix = import_optional("csr_matrix",
+                                import_from="cupyx.scipy.sparse.csr")
+cp_csc_matrix = import_optional("csc_matrix",
+                                import_from="cupyx.scipy.sparse.csc")
+
+sp = import_optional("scipy")
+sp_coo_matrix = import_optional("coo_matrix",
+                                import_from="scipy.sparse.coo")
+sp_csr_matrix = import_optional("csr_matrix",
+                                import_from="scipy.sparse.csr")
+sp_csc_matrix = import_optional("csc_matrix",
+                                import_from="scipy.sparse.csc")
 
 
-def _convert_df_to_output_type(df, input_type):
+def _ensure_args(G, source, method, directed,
+                 return_predecessors, unweighted, overwrite, indices):
+    """
+    Ensures the args passed in are usable for the API api_name and returns the
+    args with proper defaults if not specified, or raises TypeError or
+    ValueError if incorrectly specified.
+    """
+    # checks common to all input types
+    if (method is not None) and (method != "auto"):
+        raise ValueError("only 'auto' is currently accepted for method")
+    if (indices is not None) and (type(indices) == list):
+        raise ValueError("indices currently cannot be a list-like type")
+    if (indices is not None) and (source is not None):
+        raise TypeError("cannot specify both 'source' and 'indices'")
+    if (indices is None) and (source is None):
+        raise TypeError("must specify 'source' or 'indices', but not both")
+
+    G_type = type(G)
+    # Check for Graph-type inputs
+    if (G_type in [Graph, DiGraph]) or \
+       ((nx is not None) and (G_type in [nx.Graph, nx.DiGraph])):
+        exc_value = "'%s' cannot be specified for a Graph-type input"
+        if directed is not None:
+            raise TypeError(exc_value % "directed")
+        if return_predecessors is not None:
+            raise TypeError(exc_value % "return_predecessors")
+        if unweighted is not None:
+            raise TypeError(exc_value % "unweighted")
+        if overwrite is not None:
+            raise TypeError(exc_value % "overwrite")
+
+        directed = False
+
+    # Check for non-Graph-type inputs
+    else:
+        if (directed is not None) and (type(directed) != bool):
+            raise ValueError("'directed' must be a bool")
+        if (return_predecessors is not None) and \
+           (type(return_predecessors) != bool):
+            raise ValueError("'return_predecessors' must be a bool")
+        if (unweighted is not None) and (unweighted is not True):
+            raise ValueError("'unweighted' currently must be True if "
+                             "specified")
+        if (overwrite is not None) and (overwrite is not False):
+            raise ValueError("'overwrite' currently must be False if "
+                             "specified")
+
+    source = source if source is not None else indices
+    if return_predecessors is None:
+        return_predecessors = True
+
+    return (source, directed, return_predecessors)
+
+
+def _convert_df_to_output_type(df, input_type, return_predecessors):
     """
     Given a cudf.DataFrame df, convert it to a new type appropriate for the
     graph algos in this module, based on input_type.
+
+    return_predecessors is only used for return values from cupy/scipy input
+    types.
     """
     if input_type in [Graph, DiGraph]:
         return df
@@ -43,15 +110,23 @@ def _convert_df_to_output_type(df, input_type):
     elif (nx is not None) and (input_type in [nx.Graph, nx.DiGraph]):
         return df.to_pandas()
 
-    elif (cp is not None) and \
-         (input_type in [cp_coo_matrix, cp_csr_matrix, cp_csc_matrix]):
+    elif is_matrix_type(input_type):
         # A CuPy/SciPy input means the return value will be a 2-tuple of:
         #   distance: cupy.ndarray
         #   predecessor: cupy.ndarray
         sorted_df = df.sort_values("vertex")
-        return (cp.fromDlpack(sorted_df["distance"].to_dlpack()),
-                cp.fromDlpack(sorted_df["predecessor"].to_dlpack()))
-
+        if return_predecessors:
+            if is_cp_matrix_type(input_type):
+                return (cp.fromDlpack(sorted_df["distance"].to_dlpack()),
+                        cp.fromDlpack(sorted_df["predecessor"].to_dlpack()))
+            else:
+                return (sorted_df["distance"].to_array(),
+                        sorted_df["predecessor"].to_array())
+        else:
+            if is_cp_matrix_type(input_type):
+                return cp.fromDlpack(sorted_df["distance"].to_dlpack())
+            else:
+                return sorted_df["distance"].to_array()
     else:
         raise TypeError(f"input type {input_type} is not a supported type.")
 
@@ -61,7 +136,14 @@ def _convert_df_to_output_type(df, input_type):
 # Nx graphs may be needed.  From the Nx docs:
 # |      Many NetworkX algorithms designed for weighted graphs use
 # |      an edge attribute (by default `weight`) to hold a numerical value.
-def sssp(G, source):
+def sssp(G,
+         source=None,
+         method=None,
+         directed=None,
+         return_predecessors=None,
+         unweighted=None,
+         overwrite=None,
+         indices=None):
     """
     Compute the distance and predecessors for shortest paths from the specified
     source to all the vertices in the graph. The distances column will store
@@ -74,9 +156,10 @@ def sssp(G, source):
 
     Parameters
     ----------
-    graph : cuGraph.Graph, NetworkX.Graph, or CuPy sparse COO matrix
-        cuGraph graph descriptor with connectivity information. Edge weights,
-        if present, should be single or double precision floating point values.
+    graph : cugraph.Graph, networkx.Graph, CuPy or SciPy sparse matrix Graph or
+        matrix object, which should contain the connectivity information. Edge
+        weights, if present, should be single or double precision floating
+        point values.
     source : int
         Index of the source vertex.
 
@@ -100,12 +183,14 @@ def sssp(G, source):
        pandas.DataFrame with contents equivalent to the cudf.DataFrame
        described above.
 
-    If G is a CuPy sparse COO matrix, returns a 2-tuple of cupy.ndarray:
+    If G is a CuPy or SciPy matrix, returns:
+       a 2-tuple of CuPy ndarrays (if CuPy matrix input) or Numpy ndarrays (if
+       SciPy matrix input) representing:
 
-       distance: cupy.ndarray
+       distance: cupy or numpy ndarray
           ndarray of shortest distances between source and vertex.
 
-       predecessor: cupy.ndarray
+       predecessor: cupy or numpy ndarray
           ndarray of predecessors of a vertex on the path from source, which
           can be used to reconstruct the shortest paths.
 
@@ -117,10 +202,14 @@ def sssp(G, source):
     >>> G.from_cudf_edgelist(M, source='0', destination='1')
     >>> distances = cugraph.sssp(G, 0)
     """
+    (source, directed, return_predecessors) = _ensure_args(
+        G, source, method, directed, return_predecessors, unweighted,
+        overwrite, indices)
+
     # FIXME: allow nx_weight_attr to be specified
-    (G, input_type) = ensure_cugraph_obj(G,
-                                         nx_weight_attr="weight",
-                                         matrix_graph_type=Graph)
+    (G, input_type) = ensure_cugraph_obj(
+        G, nx_weight_attr="weight",
+        matrix_graph_type=DiGraph if directed else Graph)
 
     if G.renumbered:
         source = G.lookup_internal_vertex_id(cudf.Series([source]))[0]
@@ -132,7 +221,7 @@ def sssp(G, source):
         df = G.unrenumber(df, "predecessor")
         df["predecessor"].fillna(-1, inplace=True)
 
-    return _convert_df_to_output_type(df, input_type)
+    return _convert_df_to_output_type(df, input_type, return_predecessors)
 
 
 def filter_unreachable(df):
@@ -165,9 +254,17 @@ def filter_unreachable(df):
         raise TypeError("distance type unsupported")
 
 
-def shortest_path(G, source):
+def shortest_path(G,
+                  source=None,
+                  method=None,
+                  directed=None,
+                  return_predecessors=None,
+                  unweighted=None,
+                  overwrite=None,
+                  indices=None):
     """
     Alias for sssp(), provided for API compatibility with NetworkX. See sssp()
     for details.
     """
-    return sssp(G, source)
+    return sssp(G, source, method, directed, return_predecessors,
+                unweighted, overwrite, indices)