+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - PR #1265 Implement Hungarian Algorithm
 - PR #1274 Add generic from_edgelist() and from_adjlist() APIs
 - PR #1279 Add self loop check variable in graph
+- PR #1277 SciPy sparse matrix input support for WCC, SCC, SSSP, and BFS
 
 ## Improvements
 - PR #1227 Pin cmake policies to cmake 3.17 version